requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
drf-yasg==1.21.7
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
python-decouple==3.8
django-extensions==3.2.3
dj-database-url==2.1.0
